Migrate `WireAuthenticationSpike`
