+++ WireAuthentication/Sources/WireAuthentication/Components/LoginViaEmailComponentDependency.swift
@@ -24,7 +24,7 @@ internal import WireAuthenticationLogic
 
 protocol LoginViaEmailComponentDependency: Dependency {
 
-    @MainActor var router: Router { get }
+    @MainActor var router: any Router { get }
 
 }
