+++ WireAuthentication/Sources/WireAuthentication/Needle.generated.swift
