--- WireAuthenticationSpike/WireAuthentication/WireAuthentication.docc/WireAuthentication.md
@@ -1,7 +1,0 @@
-# ``WireAuthentication``
-
-## Overview
-
-## Topics
-
-### Group
